First pass at moving services to actions
