+++ src/frame/js/actions/auth.js
@@ -1,5 +1,13 @@
+import { core } from '../utils/core';
+
 export const SET_AUTH = 'SET_AUTH';
 export const RESET_AUTH = 'RESET_AUTH';
+
+export function login(props) {
+    return (dispatch, getState) => {
+        return core(getState()).appUsers.init(props);
+    };
+}
 
 export function setAuth(props) {
     return {
