+++ src/frame/js/components/Header.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import { toggleWidget, showSettings } from '../services/app';
+import { toggleWidget, showSettings } from '../actions/app';
 import { hideSettings, hideChannelPage } from '../actions/app-state';
 import { hasChannels } from '../utils/app';
 import { bindAll } from '../utils/functions';
