--- src/frame/js/services/auth.js
@@ -1,7 +1,0 @@
-import { core } from './core';
-
-export function login(props) {
-    return (dispatch, getState) => {
-        return core(getState()).appUsers.init(props);
-    };
-}
