+++ src/frame/js/smooch.jsx
@@ -9,22 +9,17 @@ import '../stylesheets/main.less';
 
 import { store } from './store';
 
-import { setAuth, resetAuth } from './actions/auth';
+import * as authActions from './actions/auth';
 import * as userActions from './actions/user';
 import { setStripeInfo, setApp } from './actions/app';
 import { updateText } from './actions/ui';
 import { setCurrentLocation } from './actions/browser';
-import { resetConversation } from './actions/conversation';
+import { sendMessage as _sendMessage, disconnectFaye, handleConversationUpdated, resetConversation } from './actions/conversation';
 import { resetIntegrations } from './actions/integrations';
-import * as AppStateActions from './actions/app-state';
+import * as appStateActions from './actions/app-state';
+import { getAccount } from './actions/stripe';
 
-import { openWidget, closeWidget } from './services/app';
-import * as authService from './services/auth';
-import { getAccount } from './services/stripe';
-import * as userService from './services/user';
-import { sendMessage as _sendMessage, disconnectFaye, handleConversationUpdated } from './services/conversation';
-import { core } from './services/core';
-
+import { core } from './utils/core';
 import { observable, observeStore } from './utils/events';
 import { waitForPage, monitorUrlChanges, stopMonitoringUrlChanges, monitorBrowserState, stopMonitoringBrowserState, updateHostClassNames } from './utils/dom';
 import { isImageUploadSupported } from './utils/media';
@@ -46,7 +41,7 @@ let unsubscribeFromStore;
 window.addEventListener('message', ({data, origin}) => {
     if (origin === `${parent.document.location.protocol}//${parent.document.location.host}`) {
         if (data.type === 'sizeChange') {
-            store.dispatch(AppStateActions.updateWidgetSize(data.value));
+            store.dispatch(appStateActions.updateWidgetSize(data.value));
         }
     }
 }, false);
@@ -119,25 +114,25 @@ export function init(props) {
     const actions = [];
 
     if (props.soundNotificationEnabled && isAudioSupported()) {
-        actions.push(AppStateActions.enableSoundNotification());
+        actions.push(appStateActions.enableSoundNotification());
     } else {
-        actions.push(AppStateActions.disableSoundNotification());
+        actions.push(appStateActions.disableSoundNotification());
     }
 
     if (props.imageUploadEnabled && isImageUploadSupported()) {
-        actions.push(AppStateActions.enableImageUpload());
+        actions.push(appStateActions.enableImageUpload());
     } else {
-        actions.push(AppStateActions.disableImageUpload());
+        actions.push(appStateActions.disableImageUpload());
     }
 
-    actions.push(AppStateActions.setEmbedded(!!props.embedded));
+    actions.push(appStateActions.setEmbedded(!!props.embedded));
 
     if (props.customText) {
         actions.push(updateText(props.customText));
     }
 
     if (props.serviceUrl) {
-        actions.push(AppStateActions.setServerURL(props.serviceUrl));
+        actions.push(appStateActions.setServerURL(props.serviceUrl));
     }
 
     store.dispatch(batchActions(actions));
@@ -151,7 +146,7 @@ export function init(props) {
     }, onStoreChange);
 
     monitorBrowserState(store.dispatch.bind(store));
-    return login(props.userId, props.jwt, pick(props, userService.EDITABLE_PROPERTIES));
+    return login(props.userId, props.jwt, pick(props, userActions.EDITABLE_PROPERTIES));
 }
 
 export function login(userId = '', jwt, attributes) {
@@ -164,25 +159,25 @@ export function login(userId = '', jwt, attributes) {
 
     const actions = [];
     // in case those are opened;
-    actions.push(AppStateActions.hideSettings());
-    actions.push(AppStateActions.hideChannelPage());
+    actions.push(appStateActions.hideSettings());
+    actions.push(appStateActions.hideChannelPage());
 
     // in case it comes from a previous authenticated state
-    actions.push(resetAuth());
+    actions.push(authActions.resetAuth());
     actions.push(userActions.resetUser());
     actions.push(resetConversation());
     actions.push(resetIntegrations());
 
 
-    attributes = pick(attributes, userService.EDITABLE_PROPERTIES);
+    attributes = pick(attributes, userActions.EDITABLE_PROPERTIES);
 
     if (store.getState().appState.emailCaptureEnabled && attributes.email) {
-        actions.push(AppStateActions.setEmailReadonly());
+        actions.push(appStateActions.setEmailReadonly());
     } else {
-        actions.push(AppStateActions.unsetEmailReadonly());
+        actions.push(appStateActions.unsetEmailReadonly());
     }
 
-    actions.push(setAuth({
+    actions.push(authActions.setAuth({
         jwt: jwt,
         appToken
     }));
@@ -194,7 +189,7 @@ export function login(userId = '', jwt, attributes) {
     initialStoreChange = true;
 
 
-    return store.dispatch(authService.login({
+    return store.dispatch(authActions.login({
         userId: userId,
         device: {
             platform: 'web',
@@ -226,7 +221,7 @@ export function login(userId = '', jwt, attributes) {
         monitorUrlChanges(() => {
             const actions = [
                 setCurrentLocation(parent.document.location),
-                userService.updateNowViewing(getDeviceId())
+                userActions.updateNowViewing(getDeviceId())
             ];
 
             store.dispatch(batchActions(actions));
@@ -242,7 +237,7 @@ export function login(userId = '', jwt, attributes) {
             });
         }
     }).then(() => {
-        return store.dispatch(userService.immediateUpdate(attributes)).then(() => {
+        return store.dispatch(userActions.immediateUpdate(attributes)).then(() => {
             const user = store.getState().user;
             if (user.conversationStarted) {
                 return store.dispatch(handleConversationUpdated());
@@ -270,7 +265,7 @@ export function sendMessage(props) {
 }
 
 export function updateUser(props) {
-    return store.dispatch(userService.update(props)).then((response) => {
+    return store.dispatch(userActions.update(props)).then((response) => {
         if (response.appUser.conversationStarted) {
             return store.dispatch(handleConversationUpdated())
                 .then(() => {
@@ -293,7 +288,7 @@ export function getConversation() {
 }
 
 export function getUserId() {
-    return userService.getUserId(store.getState());
+    return userActions.getUserId(store.getState());
 }
 
 export function getCore() {
@@ -318,11 +313,11 @@ export function destroy() {
 }
 
 export function open() {
-    store.dispatch(openWidget());
+    store.dispatch(appStateActions.openWidget());
 }
 
 export function close() {
-    store.dispatch(closeWidget());
+    store.dispatch(appStateActions.closeWidget());
 }
 
 export function isOpened() {
