+++ test/specs/actions/app.spec.js
@@ -1,11 +1,11 @@
 import sinon from 'sinon';
 
-import { openWidget, closeWidget, toggleWidget } from '../../../src/frame/js/services/app';
+import { openWidget, closeWidget, toggleWidget } from '../../../src/frame/js/actions/app';
 import { createMockedStore } from '../../utils/redux';
 import { observable } from '../../../src/frame/js/utils/events';
 import { WIDGET_STATE } from '../../../src/frame/js/constants/app';
 
-describe('App Service', () => {
+describe('App Actions', () => {
     let mockedStore;
     let sandbox;
     let clock;
